+++ src/react/atlascode/config/configController.ts
@@ -16,6 +16,7 @@ import {
     FilterSearchResponseMessage,
     JQLOptionsResponseMessage,
     JQLSuggestionsResponseMessage,
+    SiteWithAuthInfo,
     ValidateJqlResponseMessage
 } from '../../../lib/ipc/toUI/config';
 import { ConnectionTimeout } from '../../../util/time';
@@ -141,7 +142,7 @@ export type ConfigUIAction =
     | ReducerAction<ConfigUIActionType.Loading>
     | ReducerAction<
           ConfigUIActionType.SitesUpdate,
-          { jiraSites: DetailedSiteInfo[]; bitbucketSites: DetailedSiteInfo[] }
+          { jiraSites: SiteWithAuthInfo[]; bitbucketSites: SiteWithAuthInfo[] }
       >;
 
 export type ConfigChanges = { [key: string]: any };