+++ src/react/atlascode/config/jira/jql/JQLEditDialog.tsx
@@ -61,9 +61,12 @@ export const JQLEditDialog: React.FunctionComponent<JQLEditDialogProps> = ({
     const validateJql = useJqlValidator(site);
 
     register('jql', {
-        validate: async (value: string) => {
-            const jqlErrors: JQLErrors = await validateJql(value);
-            return jqlErrors.errors.length > 0 ? jqlErrors.errors.join(',') : undefined;
+        validate: async (value?: string) => {
+            if (value) {
+                const jqlErrors: JQLErrors = await validateJql(value);
+                return jqlErrors.errors.length > 0 ? jqlErrors.errors.join(',') : undefined;
+            }
+            return undefined;
         }
     });