+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## What's New in 2.4.11
+
+### Bugs fixed
+
+-   No longer show error for certain pipeline configurations
+-   Create, Search, and Configure nodes no longer disappear from Jira sidebar
+
 ## What's New in 2.4.10
 
 ### Improvements